혼공 예제 8-3, double 추가

+++ honGongE8-3.c
@@ -20,7 +20,7 @@ int main(void)
         total += score[i];
     }
 
-    avg = total / (double)count;
+    avg = (double)total / (double)count;
 
     for (i = 0; i < count; i++)
     {
